tinder/requirements.txt
altgraph
bdist-mpkg
macholib
matplotlib
modulegraph
nltk
numpy
py2app
pymongo
pyobjc-core
pyobjc-framework-Accounts
pyobjc-framework-AddressBook
pyobjc-framework-AppleScriptKit
pyobjc-framework-AppleScriptObjC
pyobjc-framework-Automator
pyobjc-framework-CFNetwork
pyobjc-framework-Cocoa
pyobjc-framework-Collaboration
pyobjc-framework-CoreData
pyobjc-framework-CoreLocation
pyobjc-framework-CoreText
pyobjc-framework-DictionaryServices
pyobjc-framework-EventKit
pyobjc-framework-ExceptionHandling
pyobjc-framework-FSEvents
pyobjc-framework-InputMethodKit
pyobjc-framework-InstallerPlugins
pyobjc-framework-InstantMessage
pyobjc-framework-LatentSemanticMapping
pyobjc-framework-LaunchServices
pyobjc-framework-Message
pyobjc-framework-OpenDirectory
pyobjc-framework-PreferencePanes
pyobjc-framework-PubSub
pyobjc-framework-QTKit
pyobjc-framework-Quartz
pyobjc-framework-ScreenSaver
pyobjc-framework-ScriptingBridge
pyobjc-framework-SearchKit
pyobjc-framework-SystemConfiguration
pyobjc-framework-WebKit
pyOpenSSL
pyparsing
pyStatParser
python-dateutil
pytz
scipy
six
Twisted
xattr
zope.interface
